add in some more refactoring to break apart this monster
